ModelReader interface migration
