+++ deeplearning4j-scaleout/deeplearning4j-nlp/src/main/java/org/deeplearning4j/models/embeddings/wordvectors/WordVectorsImpl.java
@@ -415,147 +415,8 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      */
     @Override
     public Collection<String> wordsNearest(Collection<String> positive, Collection<String> negative, int top) {
-        // Check every word is in the model
-        for (String p : SetUtils.union(new HashSet<>(positive), new HashSet<>(negative))) {
-            if (!vocab().containsWord(p)) {
-                return new ArrayList<>();
-            }
-        }
-
-        WeightLookupTable weightLookupTable = lookupTable();
-        INDArray words = Nd4j.create(positive.size() + negative.size(), weightLookupTable.layerSize());
-        int row = 0;
-        Set<String> union = SetUtils.union(new HashSet<>(positive), new HashSet<>(negative));
-        for (String s : positive) {
-            words.putRow(row++, weightLookupTable.vector(s));
-        }
-
-        for (String s : negative) {
-            words.putRow(row++, weightLookupTable.vector(s).mul(-1));
-        }
-
-        INDArray mean = words.isMatrix() ? words.mean(0) : words;
-        // TODO this should probably be replaced with wordsNearest(mean, top)
-        if (weightLookupTable instanceof InMemoryLookupTable) {
-            InMemoryLookupTable l = (InMemoryLookupTable) weightLookupTable;
-
-            INDArray syn0 = l.getSyn0();
-            syn0.diviRowVector(syn0.norm2(0));
-
-            INDArray similarity = Transforms.unitVec(mean).mmul(syn0.transpose());
-            // We assume that syn0 is normalized.
-            // Hence, the following division is not needed anymore.
-            // distances.diviRowVector(distances.norm2(1));
-            //INDArray[] sorted = Nd4j.sortWithIndices(distances,0,false);
-            List<Double> highToLowSimList = getTopN(similarity, top + union.size());
-            List<String> ret = new ArrayList<>();
-
-            for (int i = 0; i < highToLowSimList.size(); i++) {
-                String word = vocab().wordAtIndex(highToLowSimList.get(i).intValue());
-                if (word != null && !word.equals("UNK") && !word.equals("STOP") && !union.contains(word)) {
-                    ret.add(word);
-                    if (ret.size() >= top) {
-                        break;
-                    }
-                }
-            }
-
-            return ret;
-        }
-
-        Counter<String> distances = new Counter<>();
-
-        for (String s : vocab().words()) {
-            INDArray otherVec = getWordVectorMatrix(s);
-            double sim = Transforms.cosineSim(mean, otherVec);
-            distances.incrementCount(s, sim);
-        }
-
-        distances.keepTopNKeys(top);
-        return distances.keySet();
-
+        return null;
     }
-
-    private static class ArrayComparator implements Comparator<Double[]> {
-
-        @Override
-        public int compare(Double[] o1, Double[] o2) {
-                return Double.compare(o1[0], o2[0]);
-            }
-
-    }
-
-
-    /**
-     * Get top N elements
-     *
-     * @param vec the vec to extract the top elements from
-     * @param N the number of elements to extract
-     * @return the indices and the sorted top N elements
-     */
-    private static List<Double> getTopN(INDArray vec, int N) {
-        ArrayComparator comparator = new ArrayComparator();
-        PriorityQueue<Double[]> queue = new PriorityQueue<>(vec.rows(),comparator);
-
-        for (int j = 0; j < vec.length(); j++) {
-            final Double[] pair = new Double[]{vec.getDouble(j), (double) j};
-            if (queue.size() < N) {
-                queue.add(pair);
-            } else {
-                Double[] head = queue.peek();
-                if (comparator.compare(pair, head) > 0) {
-                    queue.poll();
-                    queue.add(pair);
-                }
-            }
-        }
-
-        List<Double> lowToHighSimLst = new ArrayList<>();
-
-        while (!queue.isEmpty()) {
-            double ind = queue.poll()[1];
-            lowToHighSimLst.add(ind);
-        }
-        return Lists.reverse(lowToHighSimLst);
-    }
-
-
-    /**
-     * This method returns nearest words for target word, based on tree structure.
-     * This method is recommended to use if you're going to call for nearest words multiple times.
-     *
-     * @param word
-     * @param n
-     * @param resetTree
-     * @return
-     */
-    public Collection<String> wordsNearest(String word, int n, boolean resetTree) {
-        if (!vocab.hasToken(word)) return new ArrayList<>();
-
-        // build new tree if it wasnt created before, or resetTree == TRUE
-        if (vpTree == null || resetTree) {
-            List<DataPoint> points = new ArrayList<>();
-            for (String label: vocab.words()) {
-                points.add(new DataPoint(vocab.indexOf(label), getWordVectorMatrix(label)));
-            }
-            vpTree = new VPTree(points);
-
-        }
-        List<DataPoint> add = new ArrayList<>();
-        List<Double> distances = new ArrayList<>();
-
-        // we need n+1 to address original datapoint removal
-        vpTree.search(new DataPoint(0, getWordVectorMatrix(word)), n+1, add, distances );
-
-        Collection<String> ret = new ArrayList<>();
-        for (DataPoint e: add) {
-            String label  = vocab.wordAtIndex(e.getIndex());
-            if (!label.equals(word)) ret.add(label);
-        }
-
-        return ret;
-    }
-
 
     /**
      * Get the top n words most similar to the given word
@@ -564,26 +425,7 @@ public class WordVectorsImpl<T extends SequenceElement> implements WordVectors {
      * @return the top n words
      */
     public Collection<String> wordsNearest(String word,int n) {
-        /*
-            TODO: This is temporary solution and we should get rid of flat array scan. Probably, after VPTree implementation gets fixed
-         */
-        if (!vocab.hasToken(word)) return new ArrayList<>();
-
-        INDArray mean = getWordVectorMatrix(word);
-
-        Counter<String> distances = new Counter<>();
-
-        for (String s : vocab().words()) {
-            if (s.equals(word)) continue;
-
-            INDArray otherVec = getWordVectorMatrix(s);
-            double sim = Transforms.cosineSim(mean, otherVec);
-            distances.incrementCount(s, sim);
-        }
-
-        distances.keepTopNKeys(n-1);
-        return distances.keySet();
-//        return wordsNearest(Arrays.asList(word),new ArrayList<String>(),n);
+       return null;
     }
